# workspaces/P20260914-1728-网页版票夹管理发票/03_engineering/scaffold/contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "ApiResponseInvoiceView": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/InvoiceView"
          }
        },
        "type": "object"
      },
...
      "ApiResponsePageInvoiceView": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/PageInvoiceView"
          }
        },
        "type": "object"
      },
      "ApiResponsePageRecycleItemView": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/PageRecycleItemView"
          }
        },
        "type": "object"
      },
...
      "CreateRequest": {
        "properties": {
          "amount": {
            "minLength": 1,
            "type": "string"
          },
          "category": {
            "enum": [
              "special",
              "general"
            ],
            "type": "string"
          },
          "invoiceCode": {
            "type": "string"
          },
          "invoiceNumber": {
            "minLength": 1,
            "type": "string"
          },
          "issuedDate": {
            "minLength": 1,
            "type": "string"
          },
          "medium": {
            "enum": [
              "electronic",
              "paper"
            ],
            "type": "string"
          },
          "remark": {
            "type": "string"
          },
          "status": {
            "enum": [
              "normal",
              "voided",
              "reversed"
            ],
            "type": "string"
          },
          "taxAmount": {
            "minLength": 1,
            "type": "string"
          },
          "title": {
            "minLength": 1,
            "type": "string"
          }
        },
        "required": [
          "amount",
          "invoiceNumber",
          "issuedDate",
          "taxAmount",
          "title"
        ],
        "type": "object"
      },
...
      "InvoiceView": {
        "properties": {
          "amount": {
            "type": "string"
          },
          "attachmentCount": {
            "format": "int32",
            "type": "integer"
          },
          "category": {
            "type": "string"
          },
          "createdAt": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "invoiceCode": {
            "type": "string"
          },
          "invoiceNumber": {
            "type": "string"
          },
          "issuedDate": {
            "type": "string"
          },
          "medium": {
            "type": "string"
          },
          "remark": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "taxAmount": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "totalAmount": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "PageInvoiceView": {
        "properties": {
          "optimizeCountQuery": {
            "type": "boolean",
            "writeOnly": true
          },
          "pageNumber": {
            "format": "int64",
            "type": "integer"
          },
          "pageSize": {
            "format": "int64",
            "type": "integer"
          },
          "records": {
            "items": {
              "$ref": "#/components/schemas/InvoiceView"
            },
            "type": "array"
          },
          "totalPage": {
            "format": "int64",
            "type": "integer"
          },
          "totalRow": {
            "format": "int64",
            "type": "integer"
          }
        },
        "type": "object"
      },
      "PageRecycleItemView": {
        "properties": {
          "optimizeCountQuery": {
            "type": "boolean",
            "writeOnly": true
          },
          "pageNumber": {
            "format": "int64",
            "type": "integer"
          },
          "pageSize": {
            "format": "int64",
            "type": "integer"
          },
          "records": {
            "items": {
              "$ref": "#/components/schemas/RecycleItemView"
            },
            "type": "array"
          },
          "totalPage": {
            "format": "int64",
            "type": "integer"
          },
          "totalRow": {
            "format": "int64",
            "type": "integer"
          }
        },
        "type": "object"
      },
...
      "RecycleItemView": {
        "properties": {
          "daysLeft": {
            "format": "int64",
            "type": "integer"
          },
          "deletedAt": {
            "type": "string"
          },
          "invoice": {
            "$ref": "#/components/schemas/InvoiceView"
          }
        },
        "type": "object"
      },
      "StreamingResponseBody": {},
...
    }
  },
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/invoices": {
      "get": {
        "description": "month/title(ILIKE)/amountMin-Max(作用 totalAmount)/category/medium/status 可多值逗号分隔；条件 AND、同维 OR；invoiceNumber 默认掩码，show_sensitive=1 全号",
        "operationId": "listInvoices",
        "parameters": [
          {
            "in": "query",
            "name": "month",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "title",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "amountMin",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "amountMax",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "category",
            "required": false,
            "schema": {
              "items": {
                "type": "string"
              },
              "type": "array"
            }
          },
          {
            "in": "query",
            "name": "medium",
            "required": false,
            "schema": {
              "items": {
                "type": "string"
              },
              "type": "array"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "items": {
                "type": "string"
              },
              "type": "array"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "format": "int32",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 20,
              "format": "int32",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "show_sensitive",
            "required": false,
            "schema": {
              "default": 0,
              "format": "int32",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponsePageInvoiceView"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "列表+五维筛选+分页（F07/F08）",
        "tags": [
          "invoices"
        ]
      },
      "post": {
        "description": "totalAmount 服务端重算；issuedDate 晚于今天 422 INV_002；成功 201 返回完整资源（不脱敏）",
        "operationId": "createInvoice",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseInvoiceView"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "新增发票（F04，G1）",
        "tags": [
          "invoices"
        ]
      }
    },
    "/api/v1/invoices/export": {
      "get": {
        "description": "query 同列表筛选；预检 >5000 → 422 EXP_001；200 text/csv 流式 + BOM；内容不脱敏；文件名 票夹通导出_时间戳.csv",
        "operationId": "exportInvoices",
        "parameters": [
          {
            "in": "query",
            "name": "month",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "title",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "amountMin",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "amountMax",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "category",
            "required": false,
            "schema": {
              "items": {
                "type": "string"
              },
              "type": "array"
            }
          },
          {
            "in": "query",
            "name": "medium",
            "required": false,
            "schema": {
              "items": {
                "type": "string"
              },
              "type": "array"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "items": {
                "type": "string"
              },
              "type": "array"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/StreamingResponseBody"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "CSV 导出（F10，G3）",
        "tags": [
          "invoices"
        ]
      }
    },
    "/api/v1/invoices/{id}": {
      "delete": {
        "description": "软删除（置 deletedAt），附件随行保留；再删 404 INV_001",
        "operationId": "deleteInvoice",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "summary": "删除→回收站（F06，M2）",
        "tags": [
          "invoices"
        ]
      },
      "get": {
        "description": "完整字段不脱敏（详情页口径）；非本人/不存在 404 INV_001",
        "operationId": "getInvoiceDetail",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseInvoiceView"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "发票详情（F07）",
        "tags": [
          "invoices"
        ]
      },
      "patch": {
        "description": "部分字段合并；保留 createdAt、刷新 updatedAt；校验/totalAmount 重算同创建",
        "operationId": "patchInvoice",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseInvoiceView"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "编辑发票（F05，M2）",
        "tags": [
          "invoices"
        ]
      }
    },
...
    "/api/v1/recycle/items": {
      "get": {
        "operationId": "listRecycle",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "format": "int32",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 20,
              "format": "int32",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponsePageRecycleItemView"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "回收站列表（daysLeft 服务端计算）",
        "tags": [
          "recycle"
        ]
      }
    },
    "/api/v1/recycle/items/{id}": {
      "delete": {
        "operationId": "purgeRecycleItem",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "summary": "彻底删除（级联附件+文件，不可恢复）",
        "tags": [
          "recycle"
        ]
      }
    },
    "/api/v1/recycle/items/{id}/restore": {
      "post": {
        "operationId": "restoreRecycleItem",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "summary": "恢复（置空 deletedAt）",
        "tags": [
          "recycle"
        ]
      }
    },
...
  },
  "servers": [
    {
      "description": "Generated server url",
      "url": "http://localhost:8080"
    }
  ],
...
}